Correcciones para revisión y posible aceptación de PR#13
